+++ resources/js/components/insignias/insigniaComponent.vue
@@ -1,8 +1,53 @@
 <template>
-	<div>
+	<div class="col-12">
+		<!-- crear Insignia -->
+		<form v-on:submit.prevent="agregarInsignia" enctype="multipart/form-data">
+			<div class="modal fade" id="createInsignia">
+				<div class="modal-dialog">
+					<div class="modal-content">
+						<div class="modal-header">
+							<h4>Crear</h4>
+							<button type="button" class="close" data-dismiss="modal">
+								<span>&times;</span>
+							</button>
+						</div>
+						<div class="modal-body pb-0">
+							<div class="form-group row">
+								<label class="col-form-label col-md-2">Nombre</label>
+								<div class="col-md-10">
+									<input type="text"  name="name" class="form-control" v-model="insignia.name" placeholder="Nombre de la insignia" required>
+								</div>
+							</div>
+							<div class="form-group row">
+								<label class="col-form-label col-md-2">Descripcion</label>
+								<div class="col-md-10">
+									<textarea rows="3" cols="5" name="description" class="form-control" v-model="insignia.description" placeholder="Describa la insignia"maxlength="150"  required></textarea>
+								</div>
+							</div>
+							<div class="form-group">
+								<div class="form-group">
+									<label for="imagen"></label>
+									<input type="file" @change="obtenerImagencrear" class="form-control-file" ref="img" accept="image/*" required>
+								</div>
+								<figure>
+									<img width="200" height="200" :src="imagen">
+								</figure>
+							</div>
+						</div>
+						<div class="modal-footer pb-0">
+							<div class="form-group row">
+								<div class="col-lg-6">
+									<input type="submit" name="enviar" value="Crear Registro" class="btn btn-primary pull-right">
+								</div>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</form>
 		<!-- editar registro -->
 		<form v-on:submit.prevent="updateInsignia(fillInsignia)" enctype="multipart/form-data">
-			<div class="modal fade" id="edit">
+			<div class="modal fade" id="editInsignia">
 				<div class="modal-dialog">
 					<div class="modal-content">
 						<div class="modal-header">
@@ -15,20 +60,20 @@
 							<div class="form-group row">
 								<label class="col-form-label col-md-2">Nombre</label>
 								<div class="col-md-10">
-									<input type="text"  name="name" class="form-control" v-model="fillInsignia.name">
+									<input type="text"  name="name" class="form-control" v-model="fillInsignia.name" required>
 								</div>
 							</div>
 							<div class="form-group row">
 								<label class="col-form-label col-md-2">Descripcion</label>
 								<div class="col-md-10">
-									<textarea rows="3" cols="5" name="description" class="form-control" v-model="fillInsignia.description" placeholder="Describa la insignia"></textarea>
+									<textarea rows="3" cols="5" name="description" class="form-control" v-model="fillInsignia.description" placeholder="Describa la insignia"maxlength="150"  required></textarea>
 								</div>
 							</div>
 							<div class="form-group">
 								<div class="form-group">
 									<label></label>
-									<input type="file" class="form-control-file" @change="obtenerImagen" v-if="estado == false">
-									<input type="file" class="form-control-file" @change="obtenerImagenNueva" v-else>
+									<input type="file" class="form-control-file" @change="obtenerImagen" v-if="estado == false" accept="image/*" equired>
+									<input type="file" class="form-control-file" @change="obtenerImagenNueva" v-else  accept="image/*" equired>
 								</div>
 								<figure>
 									<img width="200" height="200" :src="imagen" v-if="estado == false">
@@ -48,40 +93,39 @@
 			</div>
 		</form>
 		<!-- listar las insignias -->
-		<div class="row">
-			<div class="col-md-12">
-				<h1 class="page-header">Insignias</h1>
+		<div class="card">
+			<div class="card-header">
+				<h3 class="card-title">Insignias</h3>
+				<a href="#" class="btn btn-success float-right"  v-on:click="crearInsignia()"><i class="fas fa-plus"></i> Crear Nuevo</a>
 			</div>
-			<div class="col-sm-7">
-				<a href="#" class="btn btn-primary float-right" data-toggle="modal" data-target="#create">Crear Nuevo</a>
-				<spinner v-show="loading"></spinner>
-				<table class="table table-hover table-striped py-5">
-					<thead>
-						<tr>
-							<th>ID</th>
-							<th>Nombre</th>
-							<th>Descripcion</th>
-							<th>Imagen</th>
-							<th colspan="2">
-								&nbsp;
-							</th>
-						</tr>
-					</thead>
-					<tbody>
-						<tr v-for="(insignia, index) in insignias" :key="index" >
-							<td width="10px">{{ insignia.id}}</td>
-							<td>{{insignia.name}}</td>
-							<td>{{insignia.description}}</td>
-							<td><img :src="`imagenes/insignias/${insignia.icon}`" class="img-responsive" height="40" width="40"></td>
-							<td width="10px">
-								<a href="#" class="btn btn-warning float-right" @click="editarInsignia(insignia)">Editar</a>
-							</td>
-							<td width="10px">
-								<a href="#" class="btn btn-danger" v-on:click="eliminarInsignia(insignia,index)">Eliminar</a>
-							</td>
-						</tr>
-					</tbody>
-				</table>
+			<spinner v-if="loading"></spinner>
+			<div class="card-body" v-else>
+				<paginate name="insignias" :list="insignias" :per="5">
+					<table class="table table-bordered table-striped py-5">
+						<thead>
+							<tr>
+								<th>ID</th>
+								<th>Nombre</th>
+								<th>Descripcion</th>
+								<th class="text-center">Imagen</th>
+								<th class="text-center">Acciones</th>
+							</tr>
+						</thead>
+						<tbody>
+							<tr v-for="(insignia, index) in paginated('insignias')" :key="index" >
+								<td width="10px">{{ index+1}}</td>
+								<td>{{insignia.name}}</td>
+								<td>{{insignia.description}}</td>
+								<td class="text-center"><img :src="`imagenes/insignias/${insignia.icon}`" class="img-responsive" height="60" width="70"></td>
+								<td class="float-right">
+									<a href="#" class="btn btn-warning" @click="editarInsignia(insignia)"><i class="fas fa-pencil-alt"></i> Editar</a>
+									<a href="#" class="btn btn-danger" v-on:click="eliminarInsignia(insignia,index)"><i class="far fa-trash-alt"></i> Eliminar</a>
+								</td>
+							</tr>
+						</tbody>
+					</table>
+				</paginate>
+				<paginate-links for="insignias" :classes="{'ul': 'pagination', 'li': 'page-item', 'a': 'page-link'}"></paginate-links>
 			</div>
 		</div>
 	</div>
@@ -105,14 +149,29 @@
 				imagenMiniatura:'',
 				loading:true,
 				estado:false,
+				paginate:['insignias'],
 			}
 		},
 		methods:{
+			confirmarDelete(){
+				var resp = confirm("Estas seguro que deseas eliminarlo?");
+				if (resp == true) {
+					return true
+				}else{ return false }
+			},
 			mostrarInsignia:function(){
 				axios.get('insignia').then(res =>{
 					this.insignias = res.data
 					this.loading = false;
 				})
+			},
+			crearInsignia(){
+				$('#createInsignia').modal('show')
+			},
+			obtenerImagencrear(e){
+				let file = e.target.files[0];
+				this.insignia.icon = file;
+				this.cargarImagen(file);
 			},
 			obtenerImagen(e){
 				this.estado = true;
@@ -133,11 +192,39 @@
 				}
 				reader.readAsDataURL(file);
 			},
+			validarEspacios(parametro){
+				var patron = /^\s+$/;
+				if (patron.test(parametro)) {
+					return false
+				}else{return true}
+			},
+			agregarInsignia(){
+				if (this.validarEspacios(this.insignia.name)==false||this.validarEspacios(this.insignia.description)==false||this.validarEspacios(this.insignia.icon)==false) {
+					alert("los campos no pueden estar vacios")
+				}else{
+					let formData = new FormData();
+					formData.append('name', this.insignia.name);
+					formData.append('description', this.insignia.description);
+					formData.append('icon', this.insignia.icon);
+
+					axios.post('/insignia',formData)
+					.then(res=>{
+						EventBus.$emit('agregado', res.data.insignia);
+						this.insignia.name = "";
+						this.insignia.description = "";
+						this.$refs.img.value = "";
+						this.imagenMiniatura = "";
+						$('#createInsignia').modal('hide')
+					})
+				}
+			},
 			eliminarInsignia:function(insignia,index){
-				axios.delete(`/insignia/${insignia.id}`)
-				.then(()=>{
-					this.insignias.splice(index,1); 
-				})
+				if (this.confirmarDelete()==true) {
+					axios.delete(`/insignia/${insignia.id}`)
+					.then(()=>{
+						this.insignias.splice(index,1); 
+					})
+				}
 			},
 			editarInsignia:function (insignia){
 				this.estado = true;
@@ -145,20 +232,24 @@
 				this.fillInsignia.description = insignia.description;
 				this.fillInsignia.icon = insignia.icon;
 				this.fillInsignia.id = insignia.id;
-				$('#edit').modal('show');
+				$('#editInsignia').modal('show');
 			},
 			updateInsignia:function(fillInsignia){
-				let data = new FormData();
-				data.append('name', this.fillInsignia.name);
-				data.append('description', this.fillInsignia.description);
-				data.append('icon', this.insignia.icon);
-				data.append('_method','PUT');
-				var url = `/insignia/${fillInsignia.id}`;
-				axios.post(url, data).then(res=>{
-					this.mostrarInsignia();
-					$('#edit').modal('hide');
-				})
-			},
+				if (this.validarEspacios(this.fillInsignia.name)==false||this.validarEspacios(this.fillInsignia.description)==false||this.validarEspacios(this.fillInsignia.icon)==false) {
+					alert("los campos no pueden estar vacios")
+				}else{
+					let data = new FormData();
+					data.append('name', this.fillInsignia.name);
+					data.append('description', this.fillInsignia.description);
+					data.append('icon', this.insignia.icon);
+					data.append('_method','PUT');
+					var url = `/insignia/${fillInsignia.id}`;
+					axios.post(url, data).then(res=>{
+						this.mostrarInsignia();
+						$('#editInsignia').modal('hide');
+					})
+				}
+			}
 		},
 		computed:{
 			imagen(){
